refactor: code fixes and lint issues
